[#193] feat : 초대한 내역 서버렌더링
